Add new resolver to handle many deps at once
This is a new approach that uses a gatherer and a resolver and can
process many first-level, direct dependencies at once. It is implemented
for project-file based projects using PackageReference or legacy
Reference as well as older, non-MSBuild processable project files that
can only be parsed as plain XML.

A side effect of the many-at-once resolution is that we no longer return
tree-shaped resolved dependencies, but rather a flat list that the NuGet
libraries return.

* Add caching to NuGet gallery Catalog API requests. This is using the
  built-in caching from .Net
* Do not enable this just yet and do not update all the tests
* Add test to resolve many deps at once
* Perform minor renaming and refactoring

Reported-by: Marcel Bochtler <marcel.bochtler@bosch.com>
Reference: https://github.com/nexB/nuget-inspector/issues/23
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>